feat: Update project dependencies and configuration
This commit updates the project's dependencies to their latest stable versions for React and Vite. It also adjusts the project name and version in `package.json`. Additionally, the `index.html` file is updated to use Spanish as the default language and refactors the import map to use ESM. The `vite.config.js` is added to configure Vite and the React plugin. The `_redirects` file is also included. Finally, file extensions for component imports are removed in `.tsx` files for cleaner imports.